+++ internal/models/structs.go
@@ -66,6 +66,8 @@ type FlightRecord struct {
 	Distance float64 `json:"distance"`
 	Track    []byte  `json:"track"`
 
+	CustomFields string `json:"custom_fields"`
+
 	// calculated
 	PrevUUID string `json:"prev_uuid"`
 	NextUUID string `json:"next_uuid"`
@@ -233,4 +235,17 @@ type Currency struct {
 	} `json:"time_frame"`
 	Comparison string `json:"comparison"`
 	Filters    string `json:"filters"`
+}
+
+// CustomField is a type for custom fields
+type CustomField struct {
+	UUID          string `json:"uuid"`
+	Name          string `json:"name"`
+	Description   string `json:"description"`
+	Type          string `json:"type"`
+	StatsFunction string `json:"stats_function"`
+	SizeXs        int    `json:"size_xs"`
+	SizeMd        int    `json:"size_md"`
+	SizeLg        int    `json:"size_lg"`
+	DisplayOrder  int    `json:"display_order"`
 }